2014-09-06 21:15 UTC+0200 Viktor Szakats (vszakats users.noreply.github.com)   * contrib/gtwvg/gtwgud.c   * contrib/gtwvg/gtwvgd.c   * contrib/gtwvw/gtwvwd.c   * src/rtl/gtwvt/gtwvt.c   * contrib/gtwvg/wvgutils.c   * contrib/gtwvg/wvgwing.c   * contrib/hbwin/win_misc.c   * contrib/hbwin/wapi_winuser.c     * HB_GTI_ICONFILE, HB_GTI_ICONRES to return a pointer type handle       instead of numeric one. [INCOMPATIBLE]     * MAKEINTRESOURCE() parameter sizes cleaned and removed unnecessary casts
  * contrib/gtwvg/paint.prg
  * contrib/gtwvg/wvgutils.c
    * wvt_DlgSetIcon() rewritten in .prg

  * contrib/gtwvg/wvtwin.ch
  * contrib/gtwvg/wvtwinlg.ch
  * contrib/hbwin/hbwin.ch
    * WIN_ICON_* constants to hbwin

  * contrib/hbwin/wapi_winuser_dlg.c
    * DialogFunc() note added about some potential cleanup in passed back parameters
    + DialogFunc()/WM_COMMAND to return pointer handle type when appropriate

  * contrib/hbwin/tests/dlg.prg
    * use WIN_LOWORD() instead of relying on a non-WAPI-like feature
      (special extra callback params)
    * cleaned off numeric empty handles

+++ contrib/gtwvg/wvgwing.c
@@ -332,7 +332,7 @@ static HBITMAP hPrepareBitmap( LPCTSTR szBitmap, UINT uiBitmap,
 
          hb_snprintf( szResname, sizeof( szResname ), "?%u", uiBitmap );
 
-         hBitmap = ( HBITMAP ) LoadImage( GetModuleHandle( NULL ), MAKEINTRESOURCE( ( WORD ) uiBitmap ),
+         hBitmap = ( HBITMAP ) LoadImage( GetModuleHandle( NULL ), MAKEINTRESOURCE( uiBitmap ),
             IMAGE_BITMAP, iExpWidth, iExpHeight, bMap3Dcolors ? LR_LOADMAP3DCOLORS : LR_DEFAULTCOLOR );
          break;
       }
